abstract user id lookup
